Refactor into sdcard and FAT32 drivers, add sdcard docs
