+++ docs/fat32.md
@@ -1,0 +1,2 @@
+# FAT32
+
